Hayoo query parsing and building changed
every input is tried to parse with the Hunt query parser,
with the signature parser and, if both fail, with a simple
words parser. So query construction never fails.
When the query contains a "->" or a "=>" only the
signature parser is used.
For queries like "Monad m" both parsers are taken,
look for word "Monad" and word "m", or search for
constraint "Monad a" in signature and subsig contexts