refactor naming for app
This change makes a qualified import of the Hayoo! `App` the
easiest choice for using the App. Qualified imports are nearly
always a good choice.

For one, it allows the reader, of the code to more or less
unambiguously know, where a function or type comes from. Furthermore,
it allows for concise names in the module itself, because name
clashing should not happen on the client side, due to the qualified
import.